chore: add useActiveFilterPills hook
