+++ src/v2/Components/SavedSearchAlert/Components/ActiveFilterPills.tsx
@@ -1,16 +1,7 @@
 import React from "react"
-import { isArray } from "lodash"
-import {
-  ArtworkFilters,
-  initialArtworkFilterState,
-  useArtworkFilterContext,
-} from "v2/Components/ArtworkFilter/ArtworkFilterContext"
-import { extractPillsFromCriteria } from "../Utils/extractPills"
 import { SavedSearchAlertPills } from "./SavedSearchAlertPills"
 import { FilterPill } from "../types"
-import { getAllowedSearchCriteria } from "../Utils/savedSearchCriteria"
-import { DEFAULT_METRIC } from "v2/Components/ArtworkFilter/Utils/metrics"
-import { usePrepareFiltersForPills } from "v2/Components/ArtworkFilter/Utils/usePrepareFiltersForPills"
+import { useActiveFilterPills } from "../useActiveFilterPills"
 
 interface ActiveFilterPillsProps {
   defaultPills?: FilterPill[]
@@ -18,36 +9,7 @@ interface ActiveFilterPillsProps {
 
 export const ActiveFilterPills: React.FC<ActiveFilterPillsProps> = props => {
   const { defaultPills = [] } = props
-  const { aggregations, setFilter } = useArtworkFilterContext()
-  const filters = usePrepareFiltersForPills()
-  const criteria = getAllowedSearchCriteria(filters ?? {})
-  const metric = filters?.metric ?? DEFAULT_METRIC
-  const filterPills = extractPillsFromCriteria({
-    criteria,
-    aggregations,
-    metric,
-  })
-  const pills = [...defaultPills, ...filterPills]
-
-  const removePill = (pill: FilterPill) => {
-    if (pill.isDefault) {
-      return
-    }
-
-    let filterValue = filters![pill.field]
-
-    if (isArray(filterValue)) {
-      filterValue = filterValue.filter(value => value !== pill.value)
-    } else {
-      filterValue = initialArtworkFilterState[pill.field]
-    }
-
-    setFilter(pill.field as keyof ArtworkFilters, filterValue)
-
-    if (pill.field === "artistIDs") {
-      setFilter("includeArtworksByFollowedArtists", false)
-    }
-  }
+  const { pills, removePill } = useActiveFilterPills(defaultPills)
 
   return <SavedSearchAlertPills items={pills} onDeletePress={removePill} />
 }
